refactor types Activity{,Source,Type}
