+++ renderer/src/components/ActivityLog.tsx
@@ -1,5 +1,5 @@
 import { FC } from 'react'
-import { ActivityEventMessage } from '../typings'
+import { Activity } from '../typings'
 import dayjs from 'dayjs'
 import { ReactComponent as WarningIcon } from '../assets/img/icons/warning.svg'
 import { ReactComponent as JobIcon } from '../assets/img/icons/job.svg'
@@ -8,7 +8,7 @@ const dateTimeFormat = new Intl.DateTimeFormat(window.navigator.language, {
   hour: 'numeric', minute: 'numeric', second: 'numeric'
 })
 
-const ActivityLogItem: FC<ActivityEventMessage> = (activity) => {
+const ActivityLogItem: FC<Activity> = (activity) => {
   const time = dateTimeFormat.format(new Date(activity.timestamp))
 
   return (
@@ -42,17 +42,17 @@ const DateSeparator: FC<DateSeparatorProps> = ({ date }) => {
 }
 
 interface ActivitiesByDate {
-  [date: string]: ActivityEventMessage[];
+  [date: string]: Activity[];
 }
 interface ActivityLogProps {
-  activities: ActivityEventMessage[];
+  activities: Activity[];
 }
 
 const ActivityLog: FC<ActivityLogProps> = ({ activities = [] }) => {
   const activitiesByDate: ActivitiesByDate =
     activities
       .sort((x, y) => y.timestamp - x.timestamp)
-      .reduce((groups: ActivitiesByDate, activity: ActivityEventMessage) => {
+      .reduce((groups: ActivitiesByDate, activity: Activity) => {
         const date = dayjs(activity.timestamp).format('YYYY-MM-DD')
         return { ...groups, [date]: groups[date] ? groups[date].concat(activity) : [activity] }
       }, {})
